+++ index.js
@@ -1,5 +1,5 @@
-// 2016 December 29
-// https://github.com/bevry/editions
 'use strict'
 
-module.exports = require('editions').requirePackage(__dirname, require)
+module.exports = {
+	extends: ['eslint:recommended', 'bevry/rules', 'bevry/adapt']
+}